Fix gocognit output when a function has generic receiver type
recvString func in gocognit returns BADRECV when the function's receiver is a
generic type.

Implement recvString in two variants in order to keep the code
compatible with go versions less than 1.18
