+++ recv_pre118.go
@@ -1,0 +1,20 @@
+//go:build !go1.18
+// +build !go1.18
+
+package gocognit
+
+import (
+	"go/ast"
+)
+
+// recvString returns a string representation of recv of the
+// form "T", "*T", or "BADRECV" (if not a proper receiver type).
+func recvString(recv ast.Expr) string {
+	switch t := recv.(type) {
+	case *ast.Ident:
+		return t.Name
+	case *ast.StarExpr:
+		return "*" + recvString(t.X)
+	}
+	return "BADRECV"
+}
